requirements.txt
wandb 
pytorch-lightning
clip-anytorch
pySmartDL
transformers
Pillow==8.4.0
matplotlib
torch_cka
test_tube
pycocotools